+++ apps/web/src/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/_components/DocumentEditor/Editor/Playground/DocumentEvaluations/shared.tsx
@@ -1,0 +1,21 @@
+import {
+  ConnectedEvaluation,
+  DocumentVersion,
+  EvaluationDto,
+  EvaluationResultDto,
+} from '@latitude-data/core/browser'
+import {
+  type ICommitContextType,
+  type IProjectContextType,
+} from '@latitude-data/web-ui'
+
+export type Props = {
+  results: Record<number, EvaluationResultDto>
+  evaluations: (EvaluationDto & { live: ConnectedEvaluation['live'] })[]
+  document: DocumentVersion
+  commit: ICommitContextType['commit']
+  project: IProjectContextType['project']
+  runCount: number
+  isLoading: boolean
+  isWaiting: boolean
+}
